Code restructured and change of 'Export-Folder' now supported
When I had changed with the Export-Files into a dedicated \source sub-folder it turned out that the implementation was not ideal. Now, changing the location is just a change at one single code line. Any export service not only exports to the correct folder but also cleans the whole Workbook folder and all its sub-folders from any - at this location - outdated Export-File (the outdated then empty when dedidicated folder is yet not removed though).