Basic configuration reworked
The registry turned out not to be the perfect place to keep basic configuration values because it is user specific whereas the CompMan configuration is computer specific. Because it is imposible to write the registry under HKLM (Local Machine) without elevated permission all data is now maintained in the Workbook istself in the sheet _Basic Configuration_. The only change in the _mConfig_ module were the procedures Value (Let/Get) and Exists which had been turned to services provided by the _mWrksht_ module - which had been updated to provide a _Value_ Let/Get and a _NameExists_ service. These services are now the same as provided by the _mFile_ and the _mReg_ Common Compüonents.